Fix warnings

+++ tools/toktx/image.hpp
@@ -204,9 +204,9 @@ struct vec3 : public vec3_base {
             b = (float)(b / (double)gc_s[i]) * 2.0f - 1.0f;
             clamp(-1.0f, 1.0f);
             base_normalize();
-            r = (std::floor((r + 1.0f) * gc_s[i] * 0.5f + 0.5f));
-            g = (std::floor((g + 1.0f) * gc_s[i] * 0.5f + 0.5f));
-            b = (std::floor((b + 1.0f) * gc_s[i] * 0.5f + 0.5f));
+            r = (std::floor((r + 1.0f) * (float)gc_s[i] * 0.5f + 0.5f));
+            g = (std::floor((g + 1.0f) * (float)gc_s[i] * 0.5f + 0.5f));
+            b = (std::floor((b + 1.0f) * (float)gc_s[i] * 0.5f + 0.5f));
             clamp(0.0f, (float)gc_s[i]);
         }
     }
